локальные правки перед cherry-pick
