+++ src/entities/post/model/types.ts
@@ -1,0 +1,6 @@
+export interface Post {
+  id: number
+  title: string
+  body: string
+  userId: number
+}
